+++ pkg/model/trace/sort.go
@@ -18,7 +18,7 @@ func SortTrace(t *tempopb.Trace) {
 			})
 		}
 		sort.Slice(b.ScopeSpans, func(i, j int) bool {
-			return compareIls(b.ScopeSpans[i], b.ScopeSpans[j])
+			return compareScopeSpans(b.ScopeSpans[i], b.ScopeSpans[j])
 		})
 	}
 	sort.Slice(t.Batches, func(i, j int) bool {
@@ -28,12 +28,12 @@ func SortTrace(t *tempopb.Trace) {
 
 func compareBatches(a *v1.ResourceSpans, b *v1.ResourceSpans) bool {
 	if len(a.ScopeSpans) > 0 && len(b.ScopeSpans) > 0 {
-		return compareIls(a.ScopeSpans[0], b.ScopeSpans[0])
+		return compareScopeSpans(a.ScopeSpans[0], b.ScopeSpans[0])
 	}
 	return false
 }
 
-func compareIls(a *v1.ScopeSpans, b *v1.ScopeSpans) bool {
+func compareScopeSpans(a *v1.ScopeSpans, b *v1.ScopeSpans) bool {
 	if len(a.Spans) > 0 && len(b.Spans) > 0 {
 		return compareSpans(a.Spans[0], b.Spans[0])
 	}